fixed camera/input bug at start of the game that sent the player flying in some random location... also cleaned up python stuff, and beginnings of fully dynamic units

+++ Entity.h
@@ -188,7 +188,7 @@ public:
 
 	int team;
 	int health;
-	int completed;	// 0 -> 100 (%)
+	float completed;	// 0 -> 100 (%)
 	
 	Vector position;
 	Vector velocity;
